LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/webvtt-cross-origin.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/media/foo.vtt?url=https://www1.web-platform.test:9443/media/foo.vtt.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/media/foo.vtt?url=https://www1.web-platform.test:9443/media/foo-no-cors.vtt&mode=cors.
CONSOLE MESSAGE: Response served by service worker is CORS while mode is same origin
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/media/foo.vtt.
CONSOLE MESSAGE: Response served by service worker is CORS while mode is same origin
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/media/foo.vtt?url=https%3A%2F%2Fweb-platform.test%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3Dhttps%253A%252F%252Fwww1.web-platform.test%253A9443%252Fmedia%252Ffoo.vtt.
CONSOLE MESSAGE: FetchEvent.respondWith received an error: TypeError: Load failed
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/media/foo.vtt?url=https%3A%2F%2Fweb-platform.test%3A9443%2Fservice-workers%2Fservice-worker%2Fresources%2Fredirect.py%3FRedirect%3Dhttps%253A%252F%252Fwww1.web-platform.test%253A9443%252Fmedia%252Ffoo-no-cors.vtt&mode=cors&credentials=same-origin.
CONSOLE MESSAGE: Response served by service worker is CORS while mode is same origin
CONSOLE MESSAGE: Cannot load https://www1.web-platform.test:9443/media/foo.vtt.
CONSOLE MESSAGE: Response served by service worker is CORS while mode is same origin
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.

PASS initialize global state
PASS same-origin text track should load
PASS cross-origin text track with no-cors request should not load
PASS cross-origin text track with rejected cors request should not load
PASS cross-origin text track with approved cors request should not load
PASS same-origin text track that redirects same-origin should load
PASS same-origin text track that redirects cross-origin should not load
PASS same-origin text track that redirects to a cross-origin text track with rejected cors should not load
PASS same-origin text track that redirects to a cross-origin text track with approved cors should not load
PASS restore global state
